Stack trace unification (#52)
This pull request modifies Monica, the Walker AST engine and the Bert byte-engine so that both engines produce the same stack trace when a Monica script or Java binding throws an error, and so that those stack traces are clear and informative.

The stack trace tests are now in the JUnit test file `JoeStackTraceTest`, which executes the same tests on both `WalkerEngine` and `BertEngine`.  Normally this sort of thing would be in the scripted test suite, but the scripted
test suite can't easily test both engines at the same time, and doing so makes reconciling their behavior much easier.  Also, there are some test cases involving native bindings that are much easier to set up in Java.

A number of bugs were fixed along the way.

* The `bert.VirtualMachine` `ipSpan()` and `unwindStack()` methods are now careful to get the line number
  information for the last instruction that executed, rather than for the next
  instruction that *would have* executed had there not been an error.  This bug led to mistaken line numbers
  in some stack traces.
* `VirtualMachine::callValue` now adds a stack frame for native callables.
* `VirtualMachine::callFromJava` now adds stack frames as `WalkerEngine` does.
* `bert.Compiler` bans `return` in static initializer blocks as `WalkerEngine` does.
* Unskipped previously skipped stack trace tests in the scripted test suite; they now pass.