Joe defaults to using the Bert byte-engine. (#53)
This pull-request revised Joe and the various `joe` tools to use 
the Bert byte-engine by default instead of the older Walker AST-walker engine.

Every `joe` tool used to process scripts now has a `-w`/`--walker` option to select the
old engine, and outputs the Joe version number and engine name when and if appropriate.

In addition, `joe test` outputs timing information and `joe run` does so if given the `-t`
option.