Universal Parser (#70)
Extracted the Walker `Parser` and its ancillary classes into `joe.parser`.  Eventually it will be used for the byte-compiler as well as Walker.

Also, added the `location()` method to all `Stmt` and `Expr` values.
Some will require additional information to be saved by the parser
in order to provide the proper byte-code location.