Scanner Priming Error (#71)
The `joe.scanner.Scanner` uses a sliding window, which needs to be initialized, or "primed", before the client retrieves any tokens.  `Scanner()` attempted to prime the window by calling `advance()`; but this means that `Scanner()` can report a scanning error in the constructor, which causes problems for `joe.parser.Parser`.

This pull request modifies `Scanner` to that clients must prime it explicitly, and modifies the clients to do so safely.

* `Scanner()` no longer calls `Scanner::advance()`
* `Scanner::advance` reports the first error found at the end of the method, after all errors in sequence have been detected.
* Clients prime the scanner explicitly, and handle errors appropriately.