Clark Byte-Code Engine (#72)
This pull-request adds the new Clark byte-code engine, which replaces
the Bert engine's single-pass compiler with a two-pass compiler
sharing a parser with the Walker engine. Clark's compiler is much simpler
and easier to read and understand than Bob's, which will make evolving
Joe much simpler.

Clark is now the default engine; Bert will receive limited maintenance in the future, and will ultimately
be removed.

Clark's implementation required a handful of changes to the Parser's AST,
and hence changes to the Walker engine, but the semantics of Joe are unchanged.