Pattern matching byte-code (#75)
This pull request adds a new `PATTERN` instruction, which 
combines a constant `Pattern` and the pattern's list of evaluated 
constants into a single `PatternValue` that can more clearly be
retained on the stack and possibly reused.  `MATCH`, `GLOLET`, 
and `LOCLET` are now simple instructions expecting to find the 
`PatternValue` on the stack.