Nero Relation Schema (#105)
This pull request adds the notion of a "relation schema" to Nero.
A `Schema` defines the `Shapes` of the facts created by a Nero script's
rules and axioms.  It can be inferred from the script, or controlled by
explicit `define` declarations.  `Nero::asNeroScript` uses this to 
produce Nero scripts that can be reloaded to produce the same set of 
`Facts`.

In addition,

* `RecordFact` is now `PairFact`.
* `Parser` accumulates a `Schema` while parsing Nero code,
  and verifies that the axioms and rule heads for each relation
  have consistent shapes.
* `ASTRuleSet` now includes a `Schema`.
* `RuleSetCompiler` includes the `Schema` in the `RuleSet`.
* `RuleSetCompiler` and `RuleEngine` create `PairFacts` correctly
  given the `Schema`.
* `Nero` can now round-trip collections of facts to and from Nero source,
  assuming that all terms can be expressed as valid Nero literals.