Remove `ASTRuleSet` (#110)
Parsing Nero code to an `ASTRuleSet` and then latter compiling that to a Nero `RuleSet` has proven to be of little value.  This pull request removes the `ASTRuleSet` and its subtypes in favor of parsing directly to a `NeroRuleSet`.

In addition, Nero axioms are converted to `Fact` values in the `RuleEngine`, rather than at compilation time.