Transient Nero relations (#113)
This pull request adds the `transient` declaration and the `define transient` keyword to Nero rule sets.  A relation that is marked transient will be retained until inference is complete, and then dropped.  This can be used to drop existing relations from the input facts, as well as to drop intermediate results.

* Added relation transience to `Schema`.
* The `NeroParser` accepts the `transient` declaration and the optional `transient` keyword in `define` declarations.
* Added `transient` support to `RuleEngine`.