Nero Update Semantics (#114)
This pull request adds update semantics to Nero, e.g.,
`A!(x, y) :- A(x), B(x, y);`.  When inference is complete, `A` is dropped and `A!` is renamed `A`.  

Non-updating rules cannot use `!` in body atoms, e.g.,
`B(x) :- A!(x, y);` is forbidden.  Use `B!(x)` instead.

* Added update semantics to `RuleEngine`.
* Added parsing for the `!` update marker to `Parser`.