Parser Tests (#120)
This pull request adds `parser.ParserTest` and `parser.PatternParserTest`, and ensures they are up to date with `parser.Parser` and `parser.PatternParser`.  The point of these tests is to trigger every parser error in the parsers and verify that the error messages are what they should be.  Maintaining these tests helps to ensure that the messages follow a consistent style, that they read well and are properly informative, and that no required checks are missed.  

* Added `PatternParserTest.`
* Moved `ParserTest` from `joe.walker` to `joe.parser`, and updated the style to current practice.
* Fleshed out the declaration and statement tests and added expression tests.
* Revised all error messages to follow current style, and improved the wording where appropriate.
* Added a `foreach` test to ensure that the loop pattern contains at least one variable.