`Tool` framework changes (#128)
This pull request revises the `com.wjduquette.joe.tools` framework so that
it is no longer necessary to use `FXTool` for JavaFX-based tools.  Instead,
`ToolInfo` now has a `isJavaFX` flag; if this is true, the tool's launcher
function is invoked using the JavaFX `Platform.startup()` method.  It is then
up to the tool to create a `Stage` if it wants one, and to call
`Platform.setImplicitExit(true)` if it wants the tool to remain running until
the last window closes.

In addition, it defines a new tool, `joe testwin`, that runs Joe tests in
the JavaFX context.