Widget Revamp 4: `ListView` changes, etc. (#131)
This pull request simplifies the use of Joe's callback wrappers, and uses them in the code for the `joe.win.ListView` Joe widget.  In addition it enables
the default exception handler for the JavaFX thread in `joe win`, and fixes
a stack overflow in `joe run --clark --debug` when Joe instances define
scripted `toString` methods.