Nero API Revision (#139)
This pull request revises the Java and Joe APIs for interacting with
Nero Datalog for improved functionality and ergonomics. The Java changes are
aimed at being able to configure an instance of `Nero` or `NeroDatabase`
with client-defined entities (e.g., `Equivalences`) in an efficient and
ergonomic way.  The Joe changes replace the preliminary Joe binding
with a more ergonomic binding that's parallel to the Java binding.

The old Joe binding has been cleaned up or removed.